[IMP] move related test from test_shopinvader_payment into shopinvader_manual_payment module
