+++ test_shopinvader_payment/tests/__init__.py
@@ -1,2 +1,1 @@
 from . import test_shopinvader_quotation_payment
-from . import test_shopinvader_payment_manual
